+++ src/main/java/in/koreatech/koin/domain/timetableV2/service/TimetableLectureServiceV2.java
@@ -1,8 +1,5 @@
 package in.koreatech.koin.domain.timetableV2.service;
 
-import static in.koreatech.koin.domain.timetableV2.dto.request.TimetableLectureCreateRequest.InnerTimeTableLectureRequest;
-import static in.koreatech.koin.domain.timetableV2.dto.request.TimetableLectureUpdateRequest.InnerTimetableLectureRequest;
-import static in.koreatech.koin.domain.timetableV2.model.TimetableFrame.getDefaultTimetableFrameName;
 import static in.koreatech.koin.domain.timetableV2.util.GradeCalculator.calculateGrades;
 import static in.koreatech.koin.domain.timetableV2.util.GradeCalculator.calculateTotalGrades;
 import static in.koreatech.koin.domain.timetableV2.validation.TimetableFrameValidate.validateUserAuthorization;
@@ -15,12 +12,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import in.koreatech.koin.domain.timetable.model.Lecture;
 import in.koreatech.koin.domain.timetable.model.Semester;
-import in.koreatech.koin.domain.timetableV2.dto.request.TimetableFrameCreateRequest;
-import in.koreatech.koin.domain.timetableV2.dto.request.TimetableFrameUpdateRequest;
 import in.koreatech.koin.domain.timetableV2.dto.request.TimetableLectureCreateRequest;
 import in.koreatech.koin.domain.timetableV2.dto.request.TimetableLectureUpdateRequest;
-import in.koreatech.koin.domain.timetableV2.dto.response.TimetableFrameResponse;
-import in.koreatech.koin.domain.timetableV2.dto.response.TimetableFrameUpdateResponse;
 import in.koreatech.koin.domain.timetableV2.dto.response.TimetableLectureResponse;
 import in.koreatech.koin.domain.timetableV2.model.TimetableFrame;
 import in.koreatech.koin.domain.timetableV2.model.TimetableLecture;
@@ -31,15 +24,12 @@ import in.koreatech.koin.domain.timetableV2.repository.TimetableLectureRepositor
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.domain.user.repository.UserRepository;
 import in.koreatech.koin.global.auth.exception.AuthorizationException;
-import in.koreatech.koin.global.concurrent.ConcurrencyGuard;
-import in.koreatech.koin.global.exception.KoinIllegalArgumentException;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class TimetableServiceV2 {
-
+public class TimetableLectureServiceV2 {
     private final LectureRepositoryV2 lectureRepositoryV2;
     private final TimetableLectureRepositoryV2 timetableLectureRepositoryV2;
     private final TimetableFrameRepositoryV2 timetableFrameRepositoryV2;
@@ -47,76 +37,11 @@ public class TimetableServiceV2 {
     private final SemesterRepositoryV2 semesterRepositoryV2;
 
     @Transactional
-    public TimetableFrameResponse createTimetablesFrame(Integer userId, TimetableFrameCreateRequest request) {
-        Semester semester = semesterRepositoryV2.getBySemester(request.semester());
-        User user = userRepository.getById(userId);
-        int currentFrameCount = timetableFrameRepositoryV2.countByUserIdAndSemesterId(userId, semester.getId());
-        boolean isMain = (currentFrameCount == 0);
-        String name = (request.timetableName() != null) ? request.timetableName() :
-            getDefaultTimetableFrameName(currentFrameCount + 1);
-        TimetableFrame timetableFrame = request.toTimetablesFrame(user, semester, name, isMain);
-        TimetableFrame savedTimetableFrame = timetableFrameRepositoryV2.save(timetableFrame);
-        return TimetableFrameResponse.from(savedTimetableFrame);
-    }
-
-    @Transactional
-    public TimetableFrameUpdateResponse updateTimetableFrame(
-        TimetableFrameUpdateRequest request, Integer timetableFrameId, Integer userId
-    ) {
-        TimetableFrame timeTableFrame = timetableFrameRepositoryV2.getById(timetableFrameId);
-        Semester semester = timeTableFrame.getSemester();
-        boolean isMain = request.isMain();
-        validateTimetableFrameUpdate(userId, isMain, semester, timeTableFrame);
-        timeTableFrame.updateTimetableFrame(semester, request.timetableName(), isMain);
-        return TimetableFrameUpdateResponse.from(timeTableFrame);
-    }
-
-    private void validateTimetableFrameUpdate(Integer userId, boolean isMain, Semester semester,
-        TimetableFrame timeTableFrame) {
-        if (isMain) {
-            cancelMainTimetable(userId, semester.getId());
-            return;
-        }
-        if (timeTableFrame.isMain()) {
-            throw new KoinIllegalArgumentException("메인 시간표는 필수입니다.");
-        }
-    }
-
-    private void cancelMainTimetable(Integer userId, Integer semesterId) {
-        TimetableFrame mainTimetableFrame = timetableFrameRepositoryV2.getMainTimetableByUserIdAndSemesterId(userId,
-            semesterId);
-        mainTimetableFrame.cancelMain();
-    }
-
-    public List<TimetableFrameResponse> getTimetablesFrame(Integer userId, String semesterRequest) {
-        Semester semester = semesterRepositoryV2.getBySemester(semesterRequest);
-        return timetableFrameRepositoryV2.findAllByUserIdAndSemesterId(userId, semester.getId()).stream()
-            .map(TimetableFrameResponse::from)
-            .toList();
-    }
-
-    /*TODO. 락 3개 찾아보기 - 신관규*/
-    @ConcurrencyGuard(lockName = "deleteFrame")
-    public void deleteTimetablesFrame(Integer userId, Integer frameId) {
-        TimetableFrame timetableFrame = timetableFrameRepositoryV2.getByIdWithLock(frameId);
-        validateUserAuthorization(timetableFrame.getUser().getId(), userId);
-
-        timetableFrameRepositoryV2.deleteById(frameId);
-        if (timetableFrame.isMain()) {
-            TimetableFrame nextFrame = timetableFrameRepositoryV2.findNextFirstTimetableFrame(userId,
-                timetableFrame.getSemester().getId());
-            if (nextFrame != null) {
-                nextFrame.updateStatusMain(true);
-            }
-        }
-    }
-
-    @Transactional
     public TimetableLectureResponse createTimetableLectures(Integer userId, TimetableLectureCreateRequest request) {
         TimetableFrame timetableFrame = timetableFrameRepositoryV2.getById(request.timetableFrameId());
         validateUserAuthorization(timetableFrame.getUser().getId(), userId);
 
-        for (InnerTimeTableLectureRequest timetableLectureRequest : request.timetableLecture()) {
+        for (TimetableLectureCreateRequest.InnerTimeTableLectureRequest timetableLectureRequest : request.timetableLecture()) {
             Lecture lecture = timetableLectureRequest.lectureId() == null ?
                 null : lectureRepositoryV2.getLectureById(timetableLectureRequest.lectureId());
             TimetableLecture timetableLecture = timetableLectureRequest.toTimetableLecture(timetableFrame, lecture);
@@ -131,7 +56,7 @@ public class TimetableServiceV2 {
         TimetableFrame timetableFrame = timetableFrameRepositoryV2.getById(request.timetableFrameId());
         validateUserAuthorization(timetableFrame.getUser().getId(), userId);
 
-        for (InnerTimetableLectureRequest timetableRequest : request.timetableLecture()) {
+        for (TimetableLectureUpdateRequest.InnerTimetableLectureRequest timetableRequest : request.timetableLecture()) {
             TimetableLecture timetableLecture = timetableLectureRepositoryV2.getById(timetableRequest.id());
             timetableLecture.update(
                 timetableRequest.classTitle(),
